Bundle: type registry, theme spacing, wrapper cleanup, scroll_view rename
- type registry: single-source-of-truth via type_registry.inc; promote
  core types, MVC views, layout strategies, actions to
  ALIAS_AND_INSTANTIATE; self-contained aggregate macros in widget_list.hh
- demo ergonomics: add `namespace ui = onyxui::simple` alias and sweep
  demos to `ui::widget` concrete form; fixes CLion parser false positives
  on template-template emplace_child
- theme spacing: store spacing_values as double so fractional LU work on
  graphical backends (0.25 LU ≈ 2 px on sdlpp); propagate through
  linear_layout / grid_layout / vbox / hbox / group_box / button_group
  with on_attached hooks that refresh layout once the theme is reachable
- grid: mutable set_spacing() to update in-place without wiping set_cell()
  assignments (fixes scroll_view SIGABRT on attach)
- group_box: separate visual border-centering offset from content clearance
  so title no longer overlaps first child on pixel backends; convert
  pixel offsets to logical units via backend_metrics
- scrollbar: render track with track_normal.foreground/background via new
  draw_rect(bounds, style, fg, bg) overload on render_context; sdlpp
  theme now draws a sunken filled channel
- scroll_view: rename forwarders (add_child/emplace_child/remove_child/
  clear_children/set_layout_strategy) to `content_*` prefix; delete old
  names as stubs so stale call sites break at compile time rather than
  silently routing to ui_element and mutating the wrapper's grid; add
  double-precision scroll_to/scroll_by overloads; sweep all call sites
  (text_view, window_content_area, demo tabs, 5 test files)
- combo_box: replace private inheritance + `into_element` static_cast
  escape hatch with honest composition — combo_box is now `public
  widget<Backend>` with a combo_box_view child; `emplace_child<combo_box>`
  works like any widget; drop `into_element` entirely
- cleanup: stop tracking stale include/test_config.hh (generated by
  CMake into build/include); gitignore the path

+++ include/onyxui/theming/theme.hh
@@ -535,19 +535,23 @@ namespace onyxui {
          * ```
          */
         struct spacing_values {
-            int none = 0;
-            int tiny = 0;
-            int small = 1;
-            int medium = 1;
-            int large = 2;
-            int xlarge = 3;
+            // Stored as `double` so themes can specify fractional logical
+            // units (e.g. 0.25 LU ≈ 2 px on an 8-px backend). The layout
+            // system carries doubles throughout, so there is no rounding
+            // loss downstream.
+            double none = 0.0;
+            double tiny = 0.0;
+            double small = 1.0;
+            double medium = 1.0;
+            double large = 2.0;
+            double xlarge = 3.0;
 
             /**
-             * @brief Resolve spacing enum to logical unit integer
+             * @brief Resolve spacing enum to logical unit value
              * @param s The semantic spacing value
-             * @return Logical unit integer value
+             * @return Logical-unit value (fractional permitted)
              */
-            [[nodiscard]] constexpr int resolve(spacing s) const noexcept {
+            [[nodiscard]] constexpr double resolve(spacing s) const noexcept {
                 switch (s) {
                     case spacing::none:   return none;
                     case spacing::tiny:   return tiny;
